+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-* Added Loading screen while stored data is loading. This fixed many errors that occurred on the hosted version
-* Added toggle for showing averages on report. Defaults to off
-* added additional information to popup when you click a report cell. It now shows the cell's amount and it's deviation from average. 
+* Improved how the donut graph displays the inner label on narrower screens
+* Reworked the edit transactions popup. Each operation has a description, there's now has a submit button, and there's an option to add a manual transaction
+* Added asterisk on yearly report to indicate the cell colors are based on extrapolated amounts
+* Fixed inconsistencies with how the "Max Graph Categories" setting was interpreted.
+* Transactions from the most recent but incomplete month are now considered within the yearly report. This fixes an issue where the displayed amount and the sum amount in the cell's popup differed. 
+* Fixed bug in the math function for computing the standard deviation of multiple sets.